safeguarded lock if neopicom is active

+++ programs/local/applications/lock.bash
@@ -2,29 +2,33 @@
 
 TIMEOUT=10 # seconds
 
-# disable picom if it is enabled
-systemctl stop --user neopicom
+# Check if picom service is running
+neopicom_is_running="$(systemctl is-active --user neopicom)"
 
-# blank screen
+# Disable picom if it is enabled
+[ "${neopicom_is_running}" == "active" ] && systemctl stop --user neopicom
+
+# Blank screen
 xset dpms force off
 
-# set dpms timeout to someting fast
+# Set dpms timeout to someting fast
 xset dpms "$TIMEOUT" "$TIMEOUT" "$TIMEOUT"
 
-# lock screen
-	XSECURELOCK_AUTH_TIMEOUT="$TIMEOUT" \
-	XSECURELOCK_BLANK_TIMEOUT=5 \
-	XSECURELOCK_BLANK_DPMS_STATE="off" \
-	XSECURELOCK_NO_COMPOSITE=1 \
-	XSECURELOCK_DIM_OVERRIDE_COMPOSITOR_DETECTION=1 \
-	XSECURELOCK_PASSWORD_PROMPT="time" \
-	XSECURELOCK_SHOW_USERNAME=0 \
-	XSECURELOCK_SHOW_HOSTNAME=0 \
-	XSECURELOCK_AUTH_SOUNDS=1 \
-	xsecurelock
+# Lock screen
+XSECURELOCK_AUTH_TIMEOUT="$TIMEOUT" \
+XSECURELOCK_BLANK_TIMEOUT=5 \
+XSECURELOCK_BLANK_DPMS_STATE="off" \
+XSECURELOCK_NO_COMPOSITE=1 \
+XSECURELOCK_DIM_OVERRIDE_COMPOSITOR_DETECTION=1 \
+XSECURELOCK_PASSWORD_PROMPT="time" \
+XSECURELOCK_SHOW_USERNAME=0 \
+XSECURELOCK_SHOW_HOSTNAME=0 \
+XSECURELOCK_AUTH_SOUNDS=1 \
+xsecurelock
 
-# restore the timeout to nothing
+# Restore the timeout to nothing
 xset dpms 0 0 0
 
-# reenable picom
-systemctl start --user neopicom
+# Reenable picom if needed
+[ "${neopicom_is_running}" == "active" ] && systemctl start --user neopicom
+
